refactor: move exercise-group serlo specific code out of editor
